requirements.txt
pandas==2.2.0
matplotlib==3.8.2
numpy==1.26.3
python-docx==1.1.0
docx2pdf==0.1.8
requests==2.31.0
